refactor: move language icons to core and rename integration config types
Move PythonSVG and TypeScriptSVG into the core Icons file and delete
LanguageIcons.tsx. Rename BaseLanguageConfig to BaseIntegrationConfig
and DocsOnlyLanguageConfig to DocsOnlyIntegrationConfig.
